requirements.txt
requests==2.32.3
python-dotenv==1.1.0
ollama==0.4.8
structlog==25.3.0